java-example/step-5/CloudFormationCLI/awssamples-ec2-importkeypair/src/test/java/com/awssamples/ec2/importkeypair/DeleteHandlerTest.java
package com.awssamples.ec2.importkeypair;

import java.time.Duration;
import software.amazon.awssdk.services.ec2.Ec2Client;
import software.amazon.awssdk.services.ec2.model.DeleteKeyPairRequest;
import software.amazon.awssdk.services.ec2.model.DeleteKeyPairResponse;
import software.amazon.awssdk.services.ec2.model.DescribeKeyPairsRequest;
import software.amazon.awssdk.services.ec2.model.DescribeKeyPairsResponse;
import software.amazon.awssdk.services.ec2.model.KeyPairInfo;
import software.amazon.cloudformation.proxy.AmazonWebServicesClientProxy;
import software.amazon.cloudformation.proxy.OperationStatus;
import software.amazon.cloudformation.proxy.ProgressEvent;
import software.amazon.cloudformation.proxy.ProxyClient;
import software.amazon.cloudformation.proxy.ResourceHandlerRequest;
import software.amazon.cloudformation.exceptions.CfnGeneralServiceException;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.api.Assertions.assertThrows;
import static org.mockito.Mockito.mock;
import static org.mockito.Mockito.times;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.atLeastOnce;
import static org.mockito.Mockito.verify;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
public class DeleteHandlerTest extends AbstractTestBase {

    @Mock
    private AmazonWebServicesClientProxy proxy;

    @Mock
    private ProxyClient<Ec2Client> proxyClient;

    @Mock
    Ec2Client sdkClient;

    @BeforeEach
    public void setup() {
        proxy = new AmazonWebServicesClientProxy(logger, MOCK_CREDENTIALS, () -> Duration.ofSeconds(600).toMillis());
        sdkClient = mock(Ec2Client.class);
        proxyClient = MOCK_PROXY(proxy, sdkClient);
    }


    @Test
    public void handleRequest_SimpleSuccess() {
        final DeleteHandler handler = new DeleteHandler();

        final ResourceModel model = ResourceModel.builder()
                .keyName(this.keyName)
                .keyFingerprint(this.keyFingerprint)
                .tags(this.tags)
                .keyPairId(this.keyPairId)
                .build();
                
        final KeyPairInfo keyPairInfo = this.getKeyPairInfo(model);

        final DescribeKeyPairsResponse describeKeyPairsResponse = DescribeKeyPairsResponse
                .builder()
                .keyPairs(keyPairInfo)
                .build();
        when(proxyClient.client().describeKeyPairs(any(DescribeKeyPairsRequest.class)))
                .thenReturn(describeKeyPairsResponse);

       

        final DeleteKeyPairResponse deleteKeyPairResponse = DeleteKeyPairResponse
                .builder()
                .build();
        when(proxyClient.client().deleteKeyPair(any(DeleteKeyPairRequest.class)))
                .thenReturn(deleteKeyPairResponse);

        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                .desiredResourceState(model)
                .build();

        final ProgressEvent<ResourceModel, CallbackContext> response = handler
                .handleRequest(proxy, request, new CallbackContext(), proxyClient, logger);

        assertThat(response).isNotNull();
        assertThat(response.getStatus()).isEqualTo(OperationStatus.SUCCESS);
        assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
        assertThat(response.getResourceModels()).isNull();
        assertThat(response.getMessage()).isNull();
        assertThat(response.getErrorCode()).isNull();
        assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
        assertThat(response.getResourceModel()).isNull();

        verify(proxyClient.client()).deleteKeyPair(any(DeleteKeyPairRequest.class));
        verify(proxyClient.client(), times(1))
               .describeKeyPairs(any(DescribeKeyPairsRequest.class));
        verify(sdkClient).deleteKeyPair(any(DeleteKeyPairRequest.class));
        verify(proxyClient.client()).deleteKeyPair(any(DeleteKeyPairRequest.class));
        verify(sdkClient, atLeastOnce()).serviceName();
    }

    @Test
    public void handleRequest_AlreadyDeleted() {
        final DeleteHandler handler = new DeleteHandler();

        final ResourceModel model = ResourceModel.builder()
                .keyName(this.keyName)
                .keyFingerprint(this.keyFingerprint)
                .tags(this.tags)
                .keyPairId(this.keyPairId)
                .build();
                
        
        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                .desiredResourceState(model)
                .build();

        assertThrows(CfnGeneralServiceException.class,
                () -> handler.handleRequest(proxy, request, new CallbackContext(), proxyClient, logger));
    }
    
    @Test
    public void handleRequest_InvalidRequest() {
        final DeleteHandler handler = new DeleteHandler();

        final ResourceModel model = ResourceModel.builder()
                .build();
        final ResourceHandlerRequest<ResourceModel> request = ResourceHandlerRequest.<ResourceModel>builder()
                .desiredResourceState(model)
                .build();

        final ProgressEvent<ResourceModel, CallbackContext> response = handler
                .handleRequest(proxy, request, new CallbackContext(), proxyClient, logger);

        assertThat(response).isNotNull();
        assertThat(response.getStatus()).isEqualTo(OperationStatus.FAILED);
        assertThat(response.getCallbackDelaySeconds()).isEqualTo(0);
        assertThat(response.getResourceModel()).isEqualTo(request.getDesiredResourceState());
        assertThat(response.getResourceModels()).isNull();
        assertThat(response.getMessage()).isNotEmpty();
        assertThat(response.getErrorCode()).isNotNull();
    }
}

